+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # 📜 Changelog
 
 
+## [1.0.0] - 2025-05-06
+### Adicionado
+- Validar senhas vazias
+- Identificar senhas fracas
+- Validar senhas fortes
+- Detectar senhas comuns
+- Identificar sequências de caracteres
+- Detectar repetições
+- Fornecer sugestões de melhoria
+- Avaliar o comprimento da senha
+- Verificar a complexidade da senha
+- Lidar com casos limite
+
 ## [0.1.0] - 2025-04-03
 ### Adicionado
 - 🚀 Primeira versão do SmartSecurityPy.